Add project name to todo list, create storage module
